[FLINK-17384][flink-dist] move HBaseUtils method to HBaseConfigurationUtil.
