+++ flink-connectors/flink-connector-hbase/src/main/java/org/apache/flink/connector/hbase/HBaseTableFactory.java
@@ -26,8 +26,8 @@ import org.apache.flink.connector.hbase.options.HBaseOptions;
 import org.apache.flink.connector.hbase.options.HBaseWriteOptions;
 import org.apache.flink.connector.hbase.sink.HBaseUpsertTableSink;
 import org.apache.flink.connector.hbase.source.HBaseTableSource;
+import org.apache.flink.connector.hbase.util.HBaseConfigurationUtil;
 import org.apache.flink.connector.hbase.util.HBaseTableSchema;
-import org.apache.flink.connector.hbase.util.HBaseUtils;
 import org.apache.flink.table.api.TableSchema;
 import org.apache.flink.table.descriptors.DescriptorProperties;
 import org.apache.flink.table.descriptors.HBaseValidator;
@@ -83,7 +83,7 @@ public class HBaseTableFactory implements StreamTableSourceFactory<Row>, StreamT
 	public StreamTableSource<Row> createStreamTableSource(Map<String, String> properties) {
 		final DescriptorProperties descriptorProperties = getValidatedProperties(properties);
 		// create default configuration from current runtime env (`hbase-site.xml` in classpath) first,
-		Configuration hbaseClientConf = HBaseUtils.getHBaseConfiguration();
+		Configuration hbaseClientConf = HBaseConfigurationUtil.getHBaseConfiguration();
 		String hbaseZk = descriptorProperties.getString(CONNECTOR_ZK_QUORUM);
 		hbaseClientConf.set(HConstants.ZOOKEEPER_QUORUM, hbaseZk);
 		descriptorProperties
